+++ src/components/Microsoft.DotNet.UpgradeAssistant.MSBuild/UpgraderMsBuildExtensions.cs
@@ -23,6 +23,7 @@ namespace Microsoft.DotNet.UpgradeAssistant
             services.AddSingleton<IComponentIdentifier, ComponentIdentifier>();
             services.AddSingleton<IPackageLoader, PackageLoader>();
             services.AddSingleton<IVersionComparer, NuGetVersionComparer>();
+            services.AddSingleton<IUpgradeStartup, NuGetCredentialsStartup>();
 
             // Instantiate the upgrade context with a func to avoid needing MSBuild types prior to MSBuild registration
             services.AddTransient<MSBuildWorkspaceUpgradeContext>();
